backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
pyjwt==2.8.0
structlog==23.2.0
httpx==0.25.2
aiofiles==23.2.1
openai>=1.14.0
google-api-python-client==2.111.0
google-auth==2.26.1
python-telegram-bot==20.7
qdrant-client==1.7.3
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
sentence-transformers==2.7.0
pypdf==3.17.4
python-docx>=1.1.0
docxtpl>=0.16.0
openpyxl==3.1.2
reportlab>=4.0.0
